move calculate_area, get_contour_areas, get_space_angles, and log_gauss to plotting_tools.py
